+++ src/database/operations/bulkUploads/loadBulkUploadBundle.ts
@@ -4,15 +4,16 @@ import type {
 	Bundle,
 	BulkUpload,
 	AuthContext,
+	KeycloakUserId,
 } from '../../../types';
 
 export const loadBulkUploadBundle = async (
 	authContext: AuthContext | undefined,
-	createdBy: number | undefined,
+	createdBy: KeycloakUserId | undefined,
 	limit: number | undefined,
 	offset: number,
 ): Promise<Bundle<BulkUpload>> => {
-	const userId = authContext?.user.id;
+	const userId = authContext?.user.keycloakUserId;
 	const isAdministrator = authContext?.role.isAdministrator;
 
 	const bundle = await loadBundle<JsonResultSet<BulkUpload>>(