+++ src/database/operations/users/loadUserBundle.ts
@@ -13,7 +13,7 @@ export const loadUserBundle = async (
 	limit: number | undefined,
 	offset: number,
 ): Promise<Bundle<User>> => {
-	const userId = authContext?.user.id;
+	const userId = authContext?.user.keycloakUserId;
 	const isAdministrator = authContext?.role.isAdministrator;
 
 	const bundle = await loadBundle<JsonResultSet<User>>(