+++ src/database/queries/bulkUploads/selectWithPagination.sql
@@ -2,13 +2,13 @@ SELECT bulk_upload_to_json(bulk_uploads.*) as "object"
 FROM bulk_uploads
 WHERE
   CASE
-    WHEN :createdBy::integer IS NULL THEN
+    WHEN :createdBy::UUID IS NULL THEN
       true
     ELSE
       bulk_uploads.created_by = :createdBy
     END
   AND CASE
-    WHEN :userId::integer IS NULL THEN
+    WHEN :userId::UUID IS NULL THEN
       true
     ELSE
       (